pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "breedspotter"
version = "0.1.0"
description = "Dog breed classification with Streamlit UI and training pipeline."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "sarkaexe"}]
keywords = ["computer-vision", "classification", "streamlit", "pytorch", "dogs"]
dependencies = [
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "timm>=1.0.8",
  "streamlit>=1.36.0",
  "pillow>=10.0.0",
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "scikit-learn>=1.4.0",
  "matplotlib>=3.8.0",
  "python-dotenv>=1.0.1",
  "pydantic>=2.6.0",
  "tqdm>=4.66.0",
  "opencv-python-headless>=4.9.0.80"
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-cov>=5.0.0", "ruff>=0.6.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
